reconfig utils/index.js to utils/app.js, moved everythin from app.js into index.js file accidentally
